Added a main setup script, a partitioning script, several helper function scripts, and documented the learning process in markdown files.
cleaned up the math functions and finished writing the test.

added documentation to the math functions.

made test script executable.

initial commit of disk.sh helper function script.

added a function to return only disk devices with their symbolic link.

made a function to show disks to the user.

added a function to select a disk from the list by its index.

added a function to query the user for a disk number.

added a function that shows users all the disks and allows them to add them to the pool.

removing old prototyping functions.

added a function to query the user to select disks for the zpool and verifies that it is correct.

added documentation to the functions in disk.sh.

adding the setup and partition helper function script.

changed the dir variable name to be more accurate.

corrected the disk.sh script imports so it can be sourced from another script. imported it into the setup.sh script to use the new disk selection functions.

removed the executable flag from both scripts. changed the call of a function so that the assigned variable becomes an array. was not using arrays correctly. changed the printing of the list of selected disks so they are each on their own line.

initial commit of partition script.

added a function to get the system memory as a float so it can be rounded up for the partition size.

fixed documentation.

initial commit of questions file.

Added some knowledge about what `mount -o bind` does.

removed variable from top level setup.sh script that was relevant before using functions to define the disks.

initial commit of partition and dataset layout blog post.

added root pool and encrypted system dataset to partition script.

Added a question about how datasets with canmount=off work when a child has it on.

added system partitioning to the partition script.

added esp formatting and mounting

added creating optional user datasets to partition script.